SFA-1391: Fix ConfigUtilities
- Fix the issue in the test config utilities where it was not converted out the json values to arrays.

+++ sugarcrm/tests/SugarTestConfigUtilities.php
@@ -60,6 +60,9 @@ class SugarTestConfigUtilities
      */
     public static function resetConfig()
     {
+        if (empty(self::$orgConfig)) {
+            return;
+        }
         // delete everything in the table
         $db = DBManagerFactory::getInstance();
         $db->query('truncate config;');
@@ -69,5 +72,6 @@ class SugarTestConfigUtilities
             }
             self::$admin->saveSetting($config['category'], $config['name'], $config['value'], $config['platform']);
         }
+        self::$orgConfig = array();
     }
 }
